Reset calendar screen commit
